requirements.txt
flask
flask-cors
gunicorn
google-api-python-client
python-dotenv
groq
